feat: responsive in assignments option
