+++ src/hooks/assignments/addAssignments/useAssignmentNavigation/index.tsx
@@ -1,9 +1,13 @@
 import { useNavigate } from "react-router-dom";
-import { useEffect, useState } from "react";
+import { useContext, useEffect, useState } from "react";
 import { IFlagAppearance, useFlag } from "@inubekit/inubekit";
 import { compareObjects } from "@utils/compareObjects";
 import { withoutBusinessUnitsMessage } from "@config/assignments/generic/withoutBusinessUnitsMessage";
 import { IUseAssignmentNavigation } from "@ptypes/hooks/IUseAssignmentNavigation";
+import { formatDate } from "@utils/date/formatDate";
+import { AuthAndData } from "@context/authAndDataProvider";
+import { ISaveDataRequest } from "@ptypes/saveData/ISaveDataRequest";
+import { addAssignmentsLabels } from "@config/assignments/assisted/addAssignmentsLabels";
 
 const useAssignmentNavigation = (props: IUseAssignmentNavigation) => {
   const {
@@ -15,10 +19,16 @@ const useAssignmentNavigation = (props: IUseAssignmentNavigation) => {
     showModal,
     officialInChargeRef,
     selectedToggle,
+    absentOfficial,
     setShowModal,
   } = props;
+
+  const { appData } = useContext(AuthAndData);
   const [canRefresh, setCanRefresh] = useState<boolean>(false);
   const [showGoBackModal, setShowGoBackModal] = useState(false);
+  const [saveData, setSaveData] = useState<ISaveDataRequest>();
+
+  const [showRequestProcessModal, setShowRequestProcessModal] = useState(false);
   const { addFlag } = useFlag();
   const navigate = useNavigate();
   useEffect(() => {
@@ -105,13 +115,57 @@ const useAssignmentNavigation = (props: IUseAssignmentNavigation) => {
     (unit) => unit.isActive === true
   );
 
+  const handleSubmitClick = () => {
+    const { dateFrom, dateTo } = formValues.reasonAndCoverage.values;
+
+    const temporaryRoles = formValues.rolesByBusinessUnits.values
+      .filter((unit) => unit?.publicCode)
+      .flatMap((unit) => {
+        const roles = unit.roles || [];
+
+        return roles
+          .filter((rol) => rol.isActive)
+          .map((rol) => ({
+            roleName: rol?.value,
+            businessUnitName: unit.publicCode,
+            businessUnitCode: unit.publicCode,
+            startDate: dateFrom,
+            endDate: dateTo,
+          }));
+      });
+
+    setSaveData({
+      applicationName: "",
+      businessManagerCode: appData.businessManager.publicCode,
+      businessUnitCode: appData.businessUnit.publicCode,
+      description: addAssignmentsLabels.descriptionSaveData,
+      entityName: "Assignments",
+      requestDate: formatDate(new Date()),
+      useCaseName: "AddAssignments",
+      configurationRequestData: {
+        nameOfAbsentStaff: absentOfficial,
+        assignmentDate: dateFrom,
+        assignmentEndDate: dateTo,
+        staffName: formValues.officialInCharge.values.official,
+        staffLastName: formValues.officialInCharge.values.official,
+        staffIdentificationNumber: "",
+        temporaryRolesByBusinessUnit: temporaryRoles,
+      },
+    });
+    setShowRequestProcessModal(!showRequestProcessModal);
+  };
+
   return {
     validateSelectedToggle,
     showGoBackModal,
+    saveData,
+    showRequestProcessModal,
+    handleSubmitClick,
     handleGoBack,
     handleOpenModal,
     handleToggleModal,
     handleCloseModal,
+    setShowRequestProcessModal,
   };
 };
