+++ src/pages/assignments/tabs/assignmentsTab/addAssignments/index.tsx
@@ -1,7 +1,7 @@
 import { useContext } from "react";
 import { useLocation } from "react-router-dom";
 import { AuthAndData } from "@context/authAndDataProvider";
-import { useAddAssignments } from "@hooks/addAssignments/useAddAssignments";
+import { useAddAssignments } from "@hooks/assignments/addAssignments/useAddAssignments";
 import { useSaveAssignments } from "@hooks/assignments/saveAssigments/useSaveAssignments";
 import { EUseCase } from "@enum/useCase";
 import { addAssignmentsSteps } from "@config/assignments/assisted/steps";
